tickertape bug fix
