+++ src/public/lib/debugging.mjs
@@ -16,7 +16,7 @@ const OFF=false; */
 // document.getElementById("msg_layer").style.visibility = "hidden";
 
 // import { DEBUG_MODE } from "../allfours";
-import { gCanvasLayer } from "./screen.mjs";
+import { gCanvasLayer } from "./screenOLD.mjs";
 // import { SETTINGS }     from "./settings.mjs";
 // export const DEBUG_MODE = SETTINGS.DEBUG_MODE;
 export const DEBUG_MODE=false;
